requirements.txt
# Azure Computer Vision
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.7.1

# Image Processing
Pillow>=10.0.0

# Environment Management
python-dotenv>=1.0.0

# Web UI
streamlit>=1.29.0
plotly>=5.18.0
pandas>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
requests>=2.31.0